+++ internal/proto5server/server_listresource_test.go
@@ -5,16 +5,13 @@ package proto5server
 
 import (
 	"context"
-	"fmt"
 	"slices"
 	"strings"
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/go-cmp/cmp/cmpopts"
-	"github.com/hashicorp/go-cty/cty/msgpack"
 	"github.com/hashicorp/terraform-plugin-framework/diag"
-	"github.com/hashicorp/terraform-plugin-framework/hcl2shim"
 	"github.com/hashicorp/terraform-plugin-framework/internal/fwserver"
 	"github.com/hashicorp/terraform-plugin-framework/internal/testing/testprovider"
 	"github.com/hashicorp/terraform-plugin-framework/list"
@@ -24,8 +21,6 @@ import (
 	resourceschema "github.com/hashicorp/terraform-plugin-framework/resource/schema"
 	"github.com/hashicorp/terraform-plugin-go/tfprotov5"
 	"github.com/hashicorp/terraform-plugin-go/tftypes"
-	sdk "github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	terraformsdk "github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 )
 
 func TestServerListResource(t *testing.T) {
@@ -280,116 +275,4 @@ func TestServerListResource(t *testing.T) {
 			}
 		})
 	}
-}
-
-// a resource type defined in SDKv2
-var sdkResource sdk.Resource = sdk.Resource{
-	Schema: map[string]*sdk.Schema{
-		"id": &sdk.Schema{
-			Type: sdk.TypeString,
-		},
-		"name": &sdk.Schema{
-			Type: sdk.TypeString,
-		},
-	},
-}
-
-func listFunc(ctx context.Context, req list.ListRequest, stream *list.ListResultsStream) {
-	panic("hats")
-}
-
-func TestServerListResourceProto5ToProto5(t *testing.T) {
-	t.Parallel()
-
-	server := func(listResource func() list.ListResource) *Server {
-		return &Server{
-			FrameworkServer: fwserver.Server{
-				Provider: &testprovider.Provider{
-					ListResourcesMethod: func(ctx context.Context) []func() list.ListResource {
-						return []func() list.ListResource{listResource}
-					},
-				},
-			},
-		}
-	}
-
-	listResource := func() list.ListResource {
-		return &testprovider.ListResource{
-			ListMethod: listFunc,
-			MetadataMethod: func(ctx context.Context, req resource.MetadataRequest, resp *resource.MetadataResponse) {
-				resp.TypeName = "test_resource"
-			},
-		}
-	}
-	aServer := server(listResource)
-
-	ctx := context.Background()
-	ctx = NewContextWithSDKResource(ctx, &sdkResource)
-	req := &tfprotov5.ListResourceRequest{
-		TypeName: "test_resource",
-	}
-
-	stream, err := aServer.ListResource(ctx, req)
-	if err != nil {
-		t.Fatalf("unexpected error returned from ListResource: %v", err)
-	}
-
-	values := slices.Collect(stream.Results)
-	if len(values) > 0 {
-		if len(values[0].Diagnostics) > 0 {
-			for _, diag := range values[0].Diagnostics {
-				t.Logf("unexpected diagnostic returned from ListResource: %v", diag)
-			}
-			t.FailNow()
-		}
-	}
-
-	if len(values) == 0 {
-		t.Fatalf("expected 1 list result; got 0 list results")
-	}
-
-	// 2: from the resource type, we can obtain an initialized ResourceData value
-	d := sdkResource.Data(&terraformsdk.InstanceState{ID: "#groot"})
-
-	// 3: the initialized ResourceData value is schema-aware
-	if err := d.Set("name", "Groot"); err != nil {
-		t.Fatalf("Error setting `name`: %v", err)
-	}
-
-	if err := d.Set("nom", "groot"); err == nil {
-		t.Fatal("False negative outcome: `nom` is not a schema attribute")
-	}
-
-	displayName := "I am Groot"
-
-	// 4: mimic SDK GRPCProviderServer.ReadResource ResourceData -> MsgPack
-	state := d.State()
-	if state == nil {
-		t.Fatal("Expected state to be non-nil")
-	}
-
-	schemaBlock := sdkResource.CoreConfigSchema()
-	if schemaBlock == nil {
-		t.Fatal("Expected schemaBlock to be non-nil")
-	}
-
-	// Copied hcl2shim wholesale for purposes of making the test pass
-	newStateVal, err := hcl2shim.HCL2ValueFromFlatmap(state.Attributes, schemaBlock.ImpliedType())
-	if err != nil {
-		t.Fatalf("Error converting state attributes to HCL2 value: %v", err)
-	}
-
-	// newStateVal = normalizeNullValues(newStateVal, stateVal, false)
-
-	pack, err := msgpack.Marshal(newStateVal, schemaBlock.ImpliedType())
-	if err != nil {
-		t.Fatalf("Error marshaling new state value to MsgPack: %v", err)
-	}
-
-	fmt.Printf("MsgPack: %s\n", pack)
-
-	// 5: construct a tfprotov5.ListResourceResult
-	listResult := tfprotov5.ListResourceResult{}
-	listResult.Resource = &tfprotov5.DynamicValue{MsgPack: pack}
-	listResult.DisplayName = displayName
 }
